el carrito FUNCIONAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA!

+++ src/view/ShopWindowController.java
@@ -5,7 +5,7 @@
  */
 package view;
 
-import static com.sun.xml.internal.ws.spi.db.BindingContextFactory.LOGGER;
+// import static com.sun.xml.internal.ws.spi.db.BindingContextFactory.LOGGER;
 import controller.Controller;
 import java.awt.Desktop;
 import java.io.EOFException;
@@ -16,6 +16,7 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.ObjectInputStream;
+import java.io.ObjectOutputStream;
 import static java.lang.String.valueOf;
 import java.net.URL;
 import java.nio.file.Files;
@@ -44,6 +45,7 @@ import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.ButtonType;
+import javafx.scene.control.ChoiceDialog;
 import javafx.scene.control.Label;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.TableColumn;
@@ -61,6 +63,7 @@ import model.Company;
 import model.DBImplementation;
 import model.Product;
 import model.Profile;
+import model.Size;
 import utilidades.MyObjectOutputStream;
 import utilidades.Utilidades;
 
@@ -118,7 +121,7 @@ public class ShopWindowController implements Initializable {
         //Hacer que Items muestre Productos de la base de datos en la vista
         this.profile = profile;
         this.cont = cont;
-        ObjectInputStream ois =null;
+        ObjectInputStream ois = null;
         boolean filend = false;
         System.out.println("Perfil: " + profile);
         System.out.println("Controller: " + cont);
@@ -126,38 +129,38 @@ public class ShopWindowController implements Initializable {
         //Hacer que Items muestre Productos de la base de datos en la vista 
         // System.out.println(cont);
         uname = profile.getName();
+        fichC = new File("Carrito" + uname + ".dat");
+
+        System.out.println(uname);
         List<Product> products = cont.findAllProducts();
         if (fichC.exists()) {
             if (fichC.getName().contains(uname)) {
                 try {
+
                     ois = new ObjectInputStream(new FileInputStream(fichC));
+                    while (!filend) {
+                        CartItem ci = (CartItem) ois.readObject();
+                        talis.add(ci);
+                    }
+
+                    ois.close();
                 } catch (FileNotFoundException e) { //Excepcion no se ha encontrado el Fichero
                     e.printStackTrace();
                 } catch (IOException e) { // Excepcion error al acceder al fichero
                     e.printStackTrace();
-                }
-                try {
-                    while (!filend) {
-                        CartItem ci= (CartItem) ois.readObject();
-                        talis.add(ci);
-                        
-                    }
-                } catch (EOFException e) {
-                    filend = true;
                 } catch (ClassNotFoundException ex) {
                     Logger.getLogger(ShopWindowController.class.getName()).log(Level.SEVERE, null, ex);
                 }
 
-                carrito = FXCollections.observableList(talis);
-                tcAmout.setCellValueFactory(new PropertyValueFactory<>("atributoClase"));
-                tcItem.setCellValueFactory(new PropertyValueFactory<>("atributoClase"));
-                tcPrice.setCellValueFactory(new PropertyValueFactory<>("atributoClase"));
-
-                CartTable.setItems(carrito);
             }
-        } else {
-            fichC = new File("Carrito" + uname + ".dat");
         }
+
+        tcAmout.setCellValueFactory(new PropertyValueFactory<>("amount"));
+        tcItem.setCellValueFactory(new PropertyValueFactory<>("productName"));
+        tcPrice.setCellValueFactory(new PropertyValueFactory<>("price"));
+        carrito = FXCollections.observableList(talis);
+        CartTable.setItems(carrito);
+
         for (Product prod : products) {
             Node card = createProductCard(prod);
             productcardList.getChildren().add(card);
@@ -166,36 +169,34 @@ public class ShopWindowController implements Initializable {
 
     @FXML
     private void emptyList(ActionEvent event) {
-    clearC();
+        clearC();
     }
-    
-    private void clearC(){
-         File fichAux = new File("mediaAux.dat");
+
+    private void clearC() {
+        //File fichAux = new File("Carrito"+uname+"aux.dat");
+        // fichAux.renameTo(fichC);
+        fichC.delete();
         talis.clear();
         carrito = FXCollections.observableArrayList(talis);
-        if (fichC.delete()) {
-            fichAux.renameTo(fichC);
-        }
         CartTable.setItems(carrito);
-    
+
     }
-    
+
     @FXML
     private void buyCart(ActionEvent event) {
         Alert alert = new Alert(AlertType.CONFIRMATION);
         alert.setTitle("Confirmation");
         alert.setHeaderText("¿Do you really want to buy this?");
-        
+
         // Opcional: Personalizar botones
         // alert.getButtonTypes().setAll(ButtonType.YES, ButtonType.NO, ButtonType.CANCEL);
-        
         alert.showAndWait().ifPresent(response -> {
             if (response == ButtonType.OK) {
                 System.out.println("Usuario confirmó");
                 // Tu lógica aquí
-                for (CartItem ci: talis){
-                 //TODO borrar stock correspondiente
-                 cont.lowerStock(ci);
+                for (CartItem ci : talis) {
+                    //TODO borrar stock correspondiente
+                    cont.lowerStock(ci);
                 }
                 clearC();
                 alert.close();
@@ -313,31 +314,75 @@ public class ShopWindowController implements Initializable {
     }
 
     private void addToCart(Product product) {
-        Alert alert = new Alert(AlertType.CONFIRMATION);
+        Size sz = null;
+
+// Crear lista de opciones (strings)
+        List<String> sizeLabels = new ArrayList<>();
+        for (Size s : product.getSizes()) {
+            sizeLabels.add(s.getLabel());
+        }
+
+// Crear ChoiceDialog
+        ChoiceDialog<String> dialog = new ChoiceDialog<>(sizeLabels.get(0), sizeLabels);
+        dialog.setTitle("Choose Size");
+        dialog.setHeaderText("What size do you want?");
+        dialog.setContentText("Select a size:");
+
+// Mostrar y obtener resultado
+        String selectedLabel = dialog.showAndWait().orElse(null);
+
+        if (selectedLabel != null) {
+            // Buscar el objeto Size correspondiente
+            for (Size s : product.getSizes()) {
+                if (s.getLabel().equals(selectedLabel)) {
+                    sz = s;
+                    System.out.println(s.getLabel());
+                    break;
+                }
+            }
+        }
+
         System.out.println("añadiendo a carrito");
-        CartItem nuevoItem = new CartItem(product, "xl");
+        CartItem nuevoItem = new CartItem(product, sz);
+
         talis.add(nuevoItem);
-        try {
-            MyObjectOutputStream moos = new MyObjectOutputStream(new FileOutputStream(fichC, true));
-            moos.writeObject(nuevoItem);
-        } catch (FileNotFoundException ex) {
-            Logger.getLogger(ShopWindowController.class.getName()).log(Level.SEVERE, null, ex);
-        } catch (IOException ex) {
-            Logger.getLogger(ShopWindowController.class.getName()).log(Level.SEVERE, null, ex);
+        if (fichC.exists()) {
+            try {
+                MyObjectOutputStream moos = new MyObjectOutputStream(new FileOutputStream(fichC, true));
+                moos.writeObject(nuevoItem);
+                moos.close();
+            } catch (FileNotFoundException ex) {
+                Logger.getLogger(ShopWindowController.class.getName()).log(Level.SEVERE, null, ex);
+            } catch (IOException ex) {
+                Logger.getLogger(ShopWindowController.class.getName()).log(Level.SEVERE, null, ex);
+            }
+        } else {
+            try {
+                ObjectOutputStream oos = new ObjectOutputStream(new FileOutputStream(fichC));
+                oos.writeObject(nuevoItem);
+                oos.close();
+            } catch (FileNotFoundException ex) {
+                Logger.getLogger(ShopWindowController.class.getName()).log(Level.SEVERE, null, ex);
+            } catch (IOException ex) {
+                Logger.getLogger(ShopWindowController.class.getName()).log(Level.SEVERE, null, ex);
+            }
         }
         carrito = FXCollections.observableArrayList(talis);
+
         CartTable.setItems(carrito);
-        System.out.println("añadido a carrito\n " + carrito);
-        System.out.println("añadido a carrito\n ");
-        
+
+        System.out.println(
+                "añadido a carrito\n " + carrito);
 
     }
 
-    public static Image loadImage(String path) {
+    public static Image
+            loadImage(String path) {
 
         // 1) Classpath resource (starts with /)
         if (path.startsWith("/")) {
-            InputStream is = Product.class.getResourceAsStream(path);
+            InputStream is = Product.class
+                    .getResourceAsStream(path);
             if (is == null) {
                 throw new IllegalArgumentException("Classpath image not found: " + path);
             }
@@ -360,19 +405,19 @@ public class ShopWindowController implements Initializable {
 
     @FXML
     private void manual_open(ActionEvent event) {
-        LOGGER.info("**ShopWindow** Opening user manual");
+        //LOGGER.info("**ShopWindow** Opening user manual");
 
         try {
             File pdf = new File("pdfs/User_Manual.pdf");
             if (!pdf.exists()) {
-                LOGGER.info("**ShopWindow** User manual file not found: pdfs/User_Manual.pdf");
+                //   LOGGER.info("**ShopWindow** User manual file not found: pdfs/User_Manual.pdf");
                 return;
             }
 
             Desktop.getDesktop().open(pdf);
-            LOGGER.info("**ShopWindow** User manual opened successfully");
+            //  LOGGER.info("**ShopWindow** User manual opened successfully");
         } catch (IOException ex) {
-            LOGGER.log(Level.SEVERE, "**ShopWindow** error opening user manual", ex);
+            //   LOGGER.log(Level.SEVERE, "**ShopWindow** error opening user manual", ex);
         }
     }
